requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==21.2.0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
docutils==0.17.1
    # via flit
flit-core==3.2.0
    # via flit
flit==3.2.0
    # via -r requirements.in
idna==2.10
    # via requests
requests==2.25.1
    # via flit
selenium==3.141.0
    # via -r requirements.in
toml==0.10.2
    # via
    #   flit
    #   flit-core
urllib3==1.26.5
    # via
    #   requests
    #   selenium
